move input.js to Components folder
